pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "graphcast"
version = "0.1"
dependencies = ["cartopy",
                "chex",
                "colabtools",
                "dask",
                "dm-haiku",
                "dm-tree",
                "gmsh",
                "jax[cuda12_local]==v0.4.16",
                "jraph",
                "matplotlib",
                "numpy",
                "pandas",
                "rtree",
                "scipy",
                "torch", 
                "trimesh",
                "typing_extensions",
                "xarray",
		"zarr"]
requires-python = ">=3.11"
authors = [{name = "DeepMind"}]
description = "GraphCast: Learning skillful medium-range global weather forecasting"
readme = {file = "README.md", content-type = "text/markdown"}
keywords = ["GraphCast", "Weather", "Prediction"]
classifiers = ["Development Status :: 3 - Alpha",
               "Intended Audience :: Science/Research",
               "License :: OSI Approved :: Apache Software License",
               "Operating System :: POSIX :: Linux",
               "Programming Language :: Python :: 3.10",
               "Programming Language :: Python :: 3.11",
               "Topic :: Scientific/Engineering :: Artificial Intelligence",
               "Topic :: Scientific/Engineering :: Atmospheric Science",
               "Topic :: Scientific/Engineering :: Physics"]

[project.optional-dependencies]
profile = ["tensorboard_plugin_profile"]
interactive = ["jupyterlab", "ipywidgets", "google-cloud-storage"]

[project.urls]
Homepage = "https://github.com/deepmind/graphcast"

[tool.setuptools]
packages = ["graphcastmodel"]
